# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Conarket",
        "description": "API RESTFul do projeto conarket",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Retornar uma lista de categorias",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Registra uma nova categoria",
                "operationId": "store",
                "requestBody": {
                    "request": "Categoria",
                    "description": "Objeto de categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocess Entity"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/categories/{categoryId}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Retorna um categoria",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Id da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualiza os dados da categoria",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Id da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Categoria",
                    "description": "Objeto de categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Exclui os dados da categoria",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Id da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
        "/cooperatives/{cooperativeId}": {
            "get": {
                "tags": [
                    "Cooperativas"
                ],
                "summary": "Retorna uma cooperativa especifica",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "cooperativeId",
                        "in": "path",
                        "description": "Id da cooperativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "CooperativeResponse",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Cooperative"
                                        },
                                        {
                                            "properties": {
                                                "phones": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Phone"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "addresses": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Address"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cooperativas"
                ],
                "summary": "Atualiza uma cooperativa",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "cooperativeId",
                        "in": "path",
                        "description": "Id da cooperativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Cooperative",
                    "description": "Objeto da cooperativa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "CooperativeRequest",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/CooperativeRequest"
                                    },
                                    {
                                        "properties": {
                                            "phones": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/components/schemas/Phone"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/Address"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Not Content"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cooperativas"
                ],
                "summary": "Exclui uma cooperativa",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "cooperativeId",
                        "in": "path",
                        "description": "Id da cooperativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Not Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Cooperativas"
                ],
                "summary": "Atualiza o DAP da cooperativa",
                "operationId": "updateDap",
                "parameters": [
                    {
                        "name": "cooperativeId",
                        "in": "path",
                        "description": "Id da cooperativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "DAP",
                    "description": "Arquivo DAP da cooperativa",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "dap_path": {
                                        "description": "Arquivo do DAP no format PDF",
                                        "type": "string",
                                        "format": "base64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Not Content"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
        "/product-carts/{productCart}": {
            "delete": {
                "tags": [
                    "Carrinhos"
                ],
                "summary": "Exclui os dados do produto no carrinho",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "productCart",
                        "in": "path",
                        "description": "Id do produto no carrinho",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Carrinhos"
                ],
                "summary": "Atualiza quantidade do produto no carrinho",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "ProductCart",
                        "in": "path",
                        "description": "Id do produto no carrinho",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Quantidade do produto no carrinho",
                    "description": "Quantidade do produto no carrinho",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductCart"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
        "/products/{productId}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Retorna um produto pelo ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Id do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualiza os dados do produto",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Id do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Produto",
                    "description": "Objeto de produto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exclui os dados do produto",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Id do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
        "/users/{userId}": {
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Excluir o usuário pelo ID",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unathorizated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Address": {
                "properties": {
                    "street": {
                        "description": "Rua do endereço",
                        "type": "string"
                    },
                    "neighborhood": {
                        "description": "Bairro do endereço",
                        "type": "string"
                    },
                    "city": {
                        "description": "Cidade do endereço",
                        "type": "string"
                    },
                    "number": {
                        "description": "Número do endereço",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "properties": {
                    "id": {
                        "description": "Id da categoria",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome da categoria",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição da categoria",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CategoryRequest": {
                "properties": {
                    "name": {
                        "description": "Nome da categoria",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição da categoria",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Cooperative": {
                "properties": {
                    "name": {
                        "description": "Nome da cooperative",
                        "type": "string"
                    },
                    "dap_path": {
                        "description": "URL do arquivo DAP",
                        "type": "string"
                    },
                    "address_id": {
                        "description": "Id do endereço da cooperativa",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CooperativeRequest": {
                "properties": {
                    "name": {
                        "description": "Nome da cooperativa",
                        "type": "string"
                    },
                    "dap_path": {
                        "description": "Arquivo do DAP no format PDF",
                        "type": "string",
                        "format": "base64"
                    },
                    "city": {
                        "description": "Cidade da cooperativa",
                        "type": "string"
                    },
                    "street": {
                        "description": "Rua da cooperativa",
                        "type": "string"
                    },
                    "neighborhood": {
                        "description": "Bairro da cooperativa",
                        "type": "string"
                    },
                    "number": {
                        "description": "Número do endereço da cooperativa",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Phone": {
                "properties": {
                    "number": {
                        "description": "Número do telefone no formato (99) 99999-9999",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ProductCart": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "amount": {
                        "description": "Quantidade do produto no carrinho",
                        "type": "number"
                    },
                    "price": {
                        "description": "Preço do produto no carrinho",
                        "type": "number"
                    },
                    "delivered_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "unit_of_measure": {
                        "description": "Unidade de medida do Produto (kg ou unit)",
                        "type": "string"
                    },
                    "product_id": {
                        "description": "Id do produto",
                        "type": "integer"
                    },
                    "cart_id": {
                        "description": "Id do carrinho",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "description": "Token JWT",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer <Token>"
            }
        }
    },
    "security": [
        []
    ]
}